requirements/tests.txt
pytest==3.4.2
pytest-variables==1.7.1
pytest-selenium==1.11.4
pytest-xdist==1.22.2
pyotp==2.2.6